google-photos-dedup/requirements.txt
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
Pillow>=10.0.0
ImageHash>=4.3.1
requests>=2.31.0
selenium>=4.15.0
tqdm>=4.66.0
